Home

MajorizationExtrema.jl

MajorizationExtrema.:≺

MajorizationExtrema.TV

MajorizationExtrema.majmax

MajorizationExtrema.majmin

MajorizationExtrema.randsimplexpt

MajorizationExtrema.randsimplexpt

MajorizationExtrema.simplexpt

MajorizationExtrema.:≺ — Method.

≺(p, q)

Returns true if q majorizes p and false otherwise.

source

MajorizationExtrema.TV — Method.

TV(p, q)

Returns the total variation distance between p and q (half the sum of the absolute value of the differences).

source

MajorizationExtrema.majmax — Method.

majmax(q, ϵ)

Returns the maximum in majorization order over the total variation ball (of probability vectors) of radius ϵ around a probability vector q.

source

MajorizationExtrema.majmin — Method.

majmin(q, ϵ)

Returns the minimum in majorization order over the total variation ball (of probability vectors) of radius ϵ around a probability vector q.

source

MajorizationExtrema.randsimplexpt — Method.

randsimplexpt(d)

Generates points uniformly at random on the standard d-1 dimensional simplex using an algorithm by Smith and Tromble.

source

MajorizationExtrema.randsimplexpt — Method.

randsimplexpt(d::Int, N::Int)

Generates rational numbers with denominator at most N uniformly at random on the d-1 dimensional simplex using an algorithm by Smith and Tromble.

source

MajorizationExtrema.simplexpt — Method.

simplexpt(unif)

Takes a vector of length d-1 of numbers between 0 and 1 and converts it a point on the standard d dimensional simplex.

source