Home

MajorizationExtrema.jl

MajorizationExtrema.:≺

MajorizationExtrema.TV

MajorizationExtrema.localbound

MajorizationExtrema.majmax

MajorizationExtrema.majmin

MajorizationExtrema.randsimplexpt

MajorizationExtrema.randsimplexpt

MajorizationExtrema.simplexpt

MajorizationExtrema.:≺ — Method.

≺(p::AbstractVector{T1}, q::AbstractVector{T2}; tol = (T1 <: AbstractFloat || T2 <: AbstractFloat) ? T2(1e-8) : zero(T2) ) where {T1, T2}

Returns true if q majorizes p and false otherwise. The keyword argument tol specifies a tolerance for the comparisons. Can be used in infix form, i.e. p ≺ q.

source

MajorizationExtrema.TV — Method.

TV(p, q)

Returns the total variation distance between p and q (half the sum of the absolute value of the differences).

source

MajorizationExtrema.localbound — Method.

localbound(f, p, ϵ)

Returns δ such that |f(p) - f(q)| <= δ for any q with TV(p, q) <= ϵ, where p and q are probability vectors. Requires f to be Schur convex or Schur concave (but does not check this condition).

source

MajorizationExtrema.majmax — Method.

majmax(q, ϵ)

Returns the maximum in majorization order over the total variation ball (of probability vectors) of radius ϵ around a probability vector q.

source

MajorizationExtrema.majmin — Method.

majmin(q, ϵ)

Returns the minimum in majorization order over the total variation ball (of probability vectors) of radius ϵ around a probability vector q.

source

MajorizationExtrema.randsimplexpt — Method.

randsimplexpt(d)

Generates points uniformly at random on the standard d-1 dimensional simplex using an algorithm by Smith and Tromble.

source

MajorizationExtrema.randsimplexpt — Method.

randsimplexpt(d, N::T) where {T <: Integer}

Generates rational numbers with denominator at most N uniformly at random on the d-1 dimensional simplex using an algorithm by Smith and Tromble.

source

MajorizationExtrema.simplexpt — Method.

simplexpt(unif)

Takes a vector of length d-1 of numbers between 0 and 1 and converts it a point on the standard d dimensional simplex.

source