Home

MajorizationExtrema.jl

MajorizationExtrema.:≺

MajorizationExtrema.:≺

MajorizationExtrema.TV

MajorizationExtrema.localbound

MajorizationExtrema.majmax

MajorizationExtrema.majmax

MajorizationExtrema.majmin

MajorizationExtrema.majmin

MajorizationExtrema.randdm

MajorizationExtrema.randsimplexpt

MajorizationExtrema.randsimplexpt

MajorizationExtrema.randunitary

MajorizationExtrema.simplexpt

MajorizationExtrema.tracedist

MajorizationExtrema.:≺ — Method.

≺(ρ::Hermitian, σ::Hermitian; tol=1e-8)

Returns true if σ majorizes ρ and false otherwise. The keyword argument tol specifies a tolerance for the comparisons. Can be used in infix form, i.e. ρ ≺ σ.

source

MajorizationExtrema.:≺ — Method.

≺(p::AbstractVector{T1}, q::AbstractVector{T2}; tol = (T1 <: AbstractFloat || T2 <: AbstractFloat) ? T2(1e-8) : zero(T2) ) where {T1, T2}

Returns true if q majorizes p and false otherwise. The keyword argument tol specifies a tolerance for the comparisons. Can be used in infix form, i.e. p ≺ q.

source

MajorizationExtrema.TV — Method.

TV(p, q)

Returns the total variation distance between p and q (half the sum of the absolute value of the differences).

source

MajorizationExtrema.localbound — Method.

localbound(f, p, ϵ)

Requires f to be Schur convex or Schur concave (but does not check this condition).

If p is a vector, it is a assumed to be a probability vector, and localbound returns δ such that |f(p) - f(q)| <= δ for any probability vector q with TV(p, q) <= ϵ.

If p is a matrix, it is assumed to be a density matrix, and localbound returns δ such that |f(p) - f(q)| <= δ for any density matrix q with tracedist(p, q) <= ϵ.

source

MajorizationExtrema.majmax — Method.

majmax(q::AbstractVector, ϵ)

Returns the maximum in majorization order over the total variation ball (of probability vectors) of radius ϵ around a probability vector q.

source

MajorizationExtrema.majmax — Method.

majmax(ρ::Hermitian, ϵ)

Returns the maximum in majorization order over the trace distance ball (of density matrices) of radius ϵ around a density matrix ρ.

source

MajorizationExtrema.majmin — Method.

majmin(q::AbstractVector, ϵ)

Returns the minimum in majorization order over the total variation ball (of probability vectors) of radius ϵ around a probability vector q.

source

MajorizationExtrema.majmin — Method.

majmin(ρ::Hermitian, ϵ)

Returns the minimum in majorization order over the trace distance ball (of density matrices) of radius ϵ around a density matrix ρ.

source

MajorizationExtrema.randdm — Method.

randdm(d)

Generates a density matrix of dimension d at random.

source

MajorizationExtrema.randsimplexpt — Method.

randsimplexpt(d)

Generates points uniformly at random on the standard d-1 dimensional simplex using an algorithm by Smith and Tromble.

source

MajorizationExtrema.randsimplexpt — Method.

randsimplexpt(d, N::T) where {T <: Integer}

Generates rational numbers with denominator at most N uniformly at random on the d-1 dimensional simplex using an algorithm by Smith and Tromble.

source

MajorizationExtrema.randunitary — Method.

randunitary(d)

Generates a unitary matrix of dimension d at random according to the Haar measure, using an algorithm described by Maris Ozols in "How to generate a random unitary matrix".

source

MajorizationExtrema.simplexpt — Method.

simplexpt(unif)

Takes a vector of length d-1 of numbers between 0 and 1 and converts it a point on the standard d dimensional simplex.

source

MajorizationExtrema.tracedist — Method.

tracedist(ρ::AbstractMatrix, σ::AbstractMatrix)

Computes the trace distance between two matrices ρ and σ.

source